Add isort to our linting toolkit (#14)
This adds isort to our CI pipeline, which checks the order and format of imports. Developers can use isort to automatically do import sorting for them, and a section on this is now in the README. We configure isort to use the "black" profile, which should prevent any disagreements on import formatting between the two tools. However, in CI, we always run black last to give it the final say on formatting.