Add an uninstall command (#19)
This command removes the install and data directories, deletes the "brainframe" group, and removes all BrainFrame Docker images from the machine. It's effectively meant as the opposite of the "install" command. However, the command does not remove Docker or Docker Compose even if the CLI was originally used to install them. It also does not uninstall the BrainFrame CLI itself.